Move uscript library to own folder
