+++ lib/test.c
@@ -8,7 +8,7 @@
 #include <assert.h>
 
 #define NUMBER_TYPE int64_t
-#include "user/uscript.c"
+#include "uscript.c"
 
 #define KNRM  "\x1B[0m"
 #define KRED  "\x1B[31m"
